Move ActiveRequests logic to separate file
